Day 1- Using API calls and writing files
This commit focuses on the inital changes that I made. First, started off with using openAI and Gemini API calls. Currently will stick with Gemini as there is a free option available, but will switch later on.

Then, I focused on parsing through the input that gemini gives and writing the code segment into a python file.

Tomorrow's goal is to try doing this in different languages as well as writing methods that take in an input.